Create DTOs, Implement JPA Auditing

+++ src/main/java/com/adria/notificationsystem/repository/NotificationSysRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationSysRepository extends JpaRepository<NotificationSys,Integer> {
+public interface NotificationSysRepository extends JpaRepository<NotificationSys,Long> {
 }
